Move office to the separate microservice
