--- application/src/main/java/by/shimakser/repository/office/OfficeRepository.java
@@ -1,7 +1,0 @@
-package by.shimakser.repository.office;
-
-import by.shimakser.model.office.Office;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OfficeRepository extends JpaRepository<Office, Long> {
-}
