+++ offices/src/main/java/by/shimakser/office/Offices.java
@@ -1,0 +1,12 @@
+package by.shimakser.office;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Offices {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Offices.class, args);
+    }
+}
