+++ offices/src/main/java/by/shimakser/office/controller/OfficeOpenCsvController.java
@@ -1,7 +1,7 @@
-package by.shimakser.controller.office;
+package by.shimakser.office.controller;
 
 import by.shimakser.dto.CSVRequest;
-import by.shimakser.service.office.OfficeService;
+import by.shimakser.office.service.OfficeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
