+++ offices/src/main/java/by/shimakser/office/converter/JSONObjectConverter.java
@@ -1,4 +1,4 @@
-package by.shimakser.converter;
+package by.shimakser.office.converter;
 
 import org.json.JSONObject;
