+++ offices/src/main/java/by/shimakser/office/converter/JsonConverter.java
@@ -1,4 +1,4 @@
-package by.shimakser.converter;
+package by.shimakser.office.converter;
 
 import com.opencsv.bean.AbstractBeanField;
 import com.opencsv.exceptions.CsvConstraintViolationException;
