+++ offices/src/main/java/by/shimakser/office/model/Contact.java
@@ -1,4 +1,4 @@
-package by.shimakser.model.office;
+package by.shimakser.office.model;
 
 import lombok.*;
