+++ offices/src/main/java/by/shimakser/office/model/Status.java
@@ -1,4 +1,4 @@
-package by.shimakser.model.office;
+package by.shimakser.office.model;
 
 public enum Status {
     IN_PROCESS, UPLOADED, NOT_LOADED;
