+++ offices/src/main/java/by/shimakser/office/repository/OfficeRepository.java
@@ -1,0 +1,9 @@
+package by.shimakser.office.repository;
+
+import by.shimakser.office.model.Office;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OfficeRepository extends JpaRepository<Office, Long> {
+}
